+++ MonitorSwitcher/Model/MonitorModels.cs
@@ -15,6 +15,7 @@ namespace WorkMonitorSwitcher.Model
         public bool MinimizeToTray { get; set; }
         public bool StartWithWindows { get; set; }
         public bool ConfirmBeforeDisable { get; set; } = true;
+        public bool AutoSaveLayoutBeforeDisable { get; set; }
         public string SelectedLayoutProfile { get; set; } = "Default";
 
     }